+++ src/HotChocolate/Core/src/Types/Internal/ExtendedTypeId.cs
@@ -1,5 +1,7 @@
 using System;
 
+#nullable enable
+
 namespace HotChocolate.Internal
 {
     public readonly struct ExtendedTypeId
